appmanage_new/requirements.txt
fastapi
uvicorn
keyring
requests
keyrings.alt
requests
GitPython
PyJWT